requirements.txt
streamlit
pytest
pylint
bandit
google-generativeai
pyyaml
jsonschema
python-dotenv
numpy
